requirements.txt
opencv-python
pillow
tkinter